Fix package name Only copy Fabber libs if they exist

+++ packaging/build.py
@@ -4,7 +4,7 @@ import shutil
 import subprocess
 import re
 
-package_name = "quantiphyse_fabber_t1"
+package_name = "quantiphyse_t1"
 fabber_models_lib = "fabber_models_t1"
 
 def update_version(name, rootdir):
@@ -44,8 +44,9 @@ def build_plugin(package_name, rootdir, distdir, platform):
     # Copy Fabber shared lib
     shlib_dir, shlib_template = get_lib_template(platform)
     LIB = os.path.join(fsldir, shlib_dir, shlib_template % fabber_models_lib)
-    print("%s -> %s" % (LIB, packagedir))
-    shutil.copy(LIB, packagedir)
+    if os.path.exists(LIB):
+        print("%s -> %s" % (LIB, packagedir))
+        shutil.copy(LIB, packagedir)
 
 pkgdir = os.path.abspath(os.path.dirname(__file__))
 rootdir = os.path.abspath(os.path.join(pkgdir, os.pardir))
